% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_restatapi_cache.R
\name{clean_restatapi_cache}
\alias{clean_restatapi_cache}
\title{Clean restatapi Cache}
\usage{
clean_restatapi_cache(cache_dir = NULL, verbose = FALSE)
}
\arguments{
\item{cache_dir}{A path to cache directory. If \code{NULL} (default)
tries to clean default temporary cache directory. Directory can also be set with \code{options(restatapi_cache_dir=...)}.}

\item{verbose}{A boolean with default \code{FALSE}, so detailed messages (for debugging) will not printed. Can be set also with
\code{options(restatapi_verbose=TRUE)}}
}
\description{
Remove all objects from the .restatapi_env except the configuration file and delete all .rds files from the default and selected cache directory.
             See \code{\link{get_eurostat_data}} for more on cache.
}
\examples{

clean_restatapi_cache() 

}
