% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restatapi_env.R
\docType{data}
\name{.restatapi_env}
\alias{.restatapi_env}
\title{Create the cache environment}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.restatapi_env
}
\description{
Create the cache environment
}
\keyword{datasets}
