% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_toc.R
\name{extract_toc}
\alias{extract_toc}
\title{Extract values from XML}
\usage{
extract_toc(ns)
}
\arguments{
\item{ns}{An XML nodeset from the XML TOC file}
}
\value{
a character vector with all the values of the nodeset.
}
\description{
Extracts the values of a node from the Eurostat XML Table of contents (TOC) file
}
\details{
It is a subfunction to use in the \code{\link{get_eurostat_toc}} function.
}
\examples{
\dontrun{
values<- extract_toc(toc)
}

}
