% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compressed_sdmx.R
\name{get_compressed_sdmx}
\alias{get_compressed_sdmx}
\title{Download and extract compressed SDMX XML}
\usage{
get_compressed_sdmx(url = NULL, verbose = FALSE)
}
\arguments{
\item{url}{an XML leaf with data series from an SDMX XML file}

\item{verbose}{A boolean with default \code{FALSE}, so detailed messages (for debugging) will not printed.
Can be set also with \code{options(restatapi_verbose=TRUE)}}
}
\value{
an xml file with SDMX tags
}
\description{
Downloads  andxtracts the data values from the SDMX XML data file
}
\details{
It is a subfunction to use in the \code{\link{get_eurostat_raw}} and the \code{\link{get_eurostat_data}}  function.
}
\examples{
url<-"http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/agr_r_milkpr.sdmx.zip"
sdmx_xml<-get_compressed_sdmx(url,verbose=TRUE)

}
