% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dsd.R
\name{extract_dsd}
\alias{extract_dsd}
\title{Extract values from XML}
\usage{
extract_dsd(concept, dsd_xml)
}
\arguments{
\item{concept}{a character vector with concept ids}

\item{dsd_xml}{an XML file with DSD content}
}
\value{
a matrix with 3 columns if the concepts has code list in the DSD file
}
\description{
Extracts the values from the XML Data Structure Definition (DSD) file
}
\details{
It is a subfunction to use in the \code{\link{get_eurostat_dsd}} function.
}
\examples{
\dontrun{
dsd_url<-"http://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_nama_10_a10_e"
dsd_xml<-xml2::read_xml(dsd_url)
values<- extract_dsd("GEO",dsd_xml)
}

}
